+++ src/Exception/FileNotExistsException.php
@@ -32,10 +32,7 @@ class FileNotExistsException extends FileException
     public function __construct(?string $message = null, int $code = 0, ?\Throwable $previous = null, ?string $path = null)
     {
         if (!$message) {
-            $message = 'File or directory does not exist';
-            if ($path) {
-                $message = sprintf('File or directory `%s` does not exist', rtr($path));
-            }
+            $message = $path ? sprintf('Filename `%s` does not exist', rtr($path)) : 'Filename does not exist';
         }
         parent::__construct($message, $code, $previous, $path);
     }