+++ src/Exception/NotWritableException.php
@@ -32,10 +32,7 @@ class NotWritableException extends FileException
     public function __construct(?string $message = null, int $code = 0, ?\Throwable $previous = null, ?string $path = null)
     {
         if (!$message) {
-            $message = 'File or directory is not writable';
-            if ($path) {
-                $message = sprintf('File or directory `%s` is not writable', rtr($path));
-            }
+            $message = $path ? sprintf('Filename `%s` is not writable', rtr($path)) : 'Filename is not writable';
         }
         parent::__construct($message, $code, $previous, $path);
     }